Bug 2027697 - Use an ContiguousEnumSerializer for PropertiesFile. r=dom-core,necko-reviewers,mccr8
Differential Revision: https://phabricator.services.mozilla.com/D290841

+++ dom/console/ConsoleReportCollector.cpp
@@ -178,7 +178,7 @@ void ConsoleReportCollector::StealConsoleReports(
 
   for (const PendingReport& report : reports) {
     aReports.AppendElement(net::ConsoleReportCollected(
-        report.mErrorFlags, report.mCategory, uint32_t(report.mPropertiesFile),
+        report.mErrorFlags, report.mCategory, report.mPropertiesFile,
         report.mSourceFileURI, report.mLineNumber, report.mColumnNumber,
         report.mMessageName, report.mStringParams));
   }
